pyproject.toml
[project]
name = "enterprise-invoice-reconciler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "dspy-ai>=3.1.3",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.0",
    "langfuse>=4.0.6",
    "langgraph>=1.1.5",
    "litellm>=1.83.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "temporalio>=1.24.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
